refactor: move subscribeTo, remove cruft
- Removes old fromX functions that have not been exported since the earliest days of 5.x
- Removes redundant tests over non-exported function
- Moves subscribeTo to where `from` is, as it is the only place it is used now
- Updates partition to just use `from`.
