+++ src/internal/observable/from.ts
@@ -1,5 +1,17 @@
+import { subscribeToArray } from '../util/subscribeToArray';
+import { subscribeToPromise } from '../util/subscribeToPromise';
+import { subscribeToIterable } from '../util/subscribeToIterable';
+import { subscribeToObservable } from '../util/subscribeToObservable';
+import { isArrayLike } from '../util/isArrayLike';
+import { isPromise } from '../util/isPromise';
+import { isObject } from '../util/isObject';
+import { iterator as Symbol_iterator } from '../symbol/iterator';
+import { observable as Symbol_observable } from '../symbol/observable';
+import { Subscription } from '../Subscription';
+import { Subscriber } from '../Subscriber';
+import { subscribeToAsyncIterable } from '../util/subscribeToAsyncIterable';
+
 import { Observable } from '../Observable';
-import { subscribeTo } from '../util/subscribeTo';
 import { ObservableInput, SchedulerLike, ObservedValueOf } from '../types';
 import { scheduled } from '../scheduled/scheduled';
 
@@ -116,3 +128,25 @@ export function from<T>(input: ObservableInput<T>, scheduler?: SchedulerLike): O
     return scheduled(input, scheduler);
   }
 }
+
+function subscribeTo<T>(result: ObservableInput<T>): (subscriber: Subscriber<T>) => Subscription | void {
+  if (result && typeof (result as any)[Symbol_observable] === 'function') {
+    return subscribeToObservable(result as any);
+  } else if (isArrayLike(result)) {
+    return subscribeToArray(result);
+  } else if (isPromise(result)) {
+    return subscribeToPromise(result);
+  } else if (result && typeof (result as any)[Symbol_iterator] === 'function') {
+    return subscribeToIterable(result as any);
+  } else if (
+    Symbol && Symbol.asyncIterator &&
+    !!result && typeof (result as any)[Symbol.asyncIterator] === 'function'
+  ) {
+    return subscribeToAsyncIterable(result as any);
+  } else {
+    const value = isObject(result) ? 'an invalid object' : `'${result}'`;
+    const msg = `You provided ${value} where a stream was expected.`
+      + ' You can provide an Observable, Promise, Array, AsyncIterable, or Iterable.';
+    throw new TypeError(msg);
+  }
+};
